IGNITE-20401 Cleanup in IgniteUtils
* Fix TODOs
* Add some usages of `copyStateTo`
