+++ modules/core/src/main/java/org/apache/ignite/internal/streamer/StreamerSubscriber.java
@@ -17,6 +17,8 @@
 
 package org.apache.ignite.internal.streamer;
 
+import static org.apache.ignite.internal.util.IgniteUtils.copyStateTo;
+
 import java.util.Collection;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
@@ -223,13 +225,7 @@ public class StreamerSubscriber<T, P> implements Subscriber<T> {
 
             var futs = pendingRequests.values().toArray(new CompletableFuture[0]);
 
-            CompletableFuture.allOf(futs).whenComplete((res, err) -> {
-                if (err != null) {
-                    completionFut.completeExceptionally(err);
-                } else {
-                    completionFut.complete(null);
-                }
-            });
+            CompletableFuture.allOf(futs).whenComplete(copyStateTo(completionFut));
         } else {
             completionFut.completeExceptionally(throwable);
         }
