+++ modules/core/src/main/java/org/apache/ignite/internal/util/IgniteUtils.java
@@ -70,7 +70,6 @@ import org.apache.ignite.internal.logger.IgniteLogger;
 import org.apache.ignite.internal.manager.IgniteComponent;
 import org.apache.ignite.internal.util.worker.IgniteWorker;
 import org.apache.ignite.lang.IgniteInternalException;
-import org.apache.ignite.lang.IgniteStringBuilder;
 import org.apache.ignite.lang.IgniteStringFormatter;
 import org.apache.ignite.lang.NodeStoppingException;
 import org.jetbrains.annotations.Nullable;
@@ -79,8 +78,6 @@ import org.jetbrains.annotations.Nullable;
  * Collection of utility methods used throughout the system.
  */
 public class IgniteUtils {
-    /** Byte bit-mask. */
-    private static final int MASK = 0xf;
 
     /** The moment will be used as a start monotonic time. */
     private static final long BEGINNING_OF_TIME = System.nanoTime();
@@ -254,125 +251,6 @@ public class IgniteUtils {
     }
 
     /**
-     * Converts byte array to hex string.
-     *
-     * @param arr Array of bytes.
-     * @return Hex string.
-     */
-    public static String toHexString(byte[] arr) {
-        return toHexString(arr, Integer.MAX_VALUE);
-    }
-
-    /**
-     * Converts byte array to hex string.
-     *
-     * @param arr Array of bytes.
-     * @param maxLen Maximum length of result string. Rounds down to a power of two.
-     * @return Hex string.
-     */
-    public static String toHexString(byte[] arr, int maxLen) {
-        assert maxLen >= 0 : "maxLem must be not negative.";
-
-        int capacity = Math.min(arr.length << 1, maxLen);
-
-        int lim = capacity >> 1;
-
-        StringBuilder sb = new StringBuilder(capacity);
-
-        for (int i = 0; i < lim; i++) {
-            addByteAsHex(sb, arr[i]);
-        }
-
-        return sb.toString().toUpperCase();
-    }
-
-    /**
-     * Returns hex representation of memory region.
-     *
-     * @param addr Pointer in memory.
-     * @param len How much byte to read.
-     */
-    public static String toHexString(long addr, int len) {
-        StringBuilder sb = new StringBuilder(len * 2);
-
-        for (int i = 0; i < len; i++) {
-            // Can not use getLong because on little-endian it produces wrong result.
-            addByteAsHex(sb, GridUnsafe.getByte(addr + i));
-        }
-
-        return sb.toString();
-    }
-
-    /**
-     * Returns hex representation of memory region.
-     *
-     * @param buf Buffer which content should be converted to string.
-     */
-    public static String toHexString(ByteBuffer buf) {
-        StringBuilder sb = new StringBuilder(buf.capacity() * 2);
-
-        for (int i = buf.position(); i < buf.limit(); i++) {
-            // Can not use getLong because on little-endian it produces wrong result.
-            addByteAsHex(sb, buf.get(i));
-        }
-
-        return sb.toString();
-    }
-
-    /**
-     * Returns byte array represented by given hex string.
-     *
-     * @param s String containing a hex representation of bytes.
-     * @return A byte array.
-     */
-    public static byte[] fromHexString(String s) {
-        var len = s.length();
-
-        assert (len & 1) == 0 : "length should be even";
-
-        var data = new byte[len / 2];
-
-        for (int i = 0; i < len; i += 2) {
-            data[i / 2] = (byte) ((Character.digit(s.charAt(i), 16) << 4)
-                    + Character.digit(s.charAt(i + 1), 16));
-        }
-
-        return data;
-    }
-
-    /**
-     * Appends {@code byte} in hexadecimal format.
-     *
-     * @param sb String builder.
-     * @param b Byte to add in hexadecimal format.
-     */
-    private static void addByteAsHex(StringBuilder sb, byte b) {
-        sb.append(Integer.toHexString(MASK & b >>> 4)).append(Integer.toHexString(MASK & b));
-    }
-
-    /**
-     * Returns a hex string representation of the given long value.
-     *
-     * @param val Value to convert to string.
-     * @return Hex string.
-     */
-    //TODO IGNITE-16350 Consider renaming or moving into other class.
-    public static String hexLong(long val) {
-        return new IgniteStringBuilder(16).appendHex(val).toString();
-    }
-
-    /**
-     * Returns a hex string representation of the given integer value.
-     *
-     * @param val Value to convert to string.
-     * @return Hex string.
-     */
-    //TODO IGNITE-16350 Consider renaming or moving into other class.
-    public static String hexInt(int val) {
-        return new IgniteStringBuilder(8).appendHex(val).toString();
-    }
-
-    /**
      * Returns size in human-readable format.
      *
      * @param bytes Number of bytes to display.
@@ -468,7 +346,7 @@ public class IgniteUtils {
     public static Class<?> forName(
             String clsName,
             @Nullable ClassLoader ldr,
-            Predicate<String> clsFilter
+            @Nullable Predicate<String> clsFilter
     ) throws ClassNotFoundException {
         assert clsName != null;
 
@@ -732,6 +610,7 @@ public class IgniteUtils {
      * @param msg Message to print with the stack.
      * @deprecated Calls to this method should never be committed to master.
      */
+    @Deprecated
     public static void dumpStack(IgniteLogger log, String msg, Object... params) {
         String reason = "Dumping stack";
 
@@ -766,7 +645,7 @@ public class IgniteUtils {
 
         try {
             success = Files.move(sourcePath, targetPath, StandardCopyOption.ATOMIC_MOVE);
-        } catch (final IOException e) {
+        } catch (IOException e) {
             // If it falls here that can mean many things. Either that the atomic move is not supported,
             // or something wrong happened. Anyway, let's try to be over-diagnosing
             if (log != null) {
@@ -783,7 +662,7 @@ public class IgniteUtils {
 
             try {
                 success = Files.move(sourcePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
-            } catch (final IOException e1) {
+            } catch (IOException e1) {
                 e1.addSuppressed(e);
 
                 if (log != null) {
@@ -795,7 +674,7 @@ public class IgniteUtils {
 
                 try {
                     Files.deleteIfExists(sourcePath);
-                } catch (final IOException e2) {
+                } catch (IOException e2) {
                     e2.addSuppressed(e1);
 
                     if (log != null) {
