+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/util/PageIdUtils.java
@@ -21,7 +21,7 @@ import static org.apache.ignite.internal.pagememory.PageIdAllocator.FLAG_DATA;
 
 import org.apache.ignite.internal.pagememory.FullPageId;
 import org.apache.ignite.internal.pagememory.PageIdAllocator;
-import org.apache.ignite.internal.util.IgniteUtils;
+import org.apache.ignite.internal.util.HexStringUtils;
 
 /**
  * Utility class for page ID parts manipulation.
@@ -88,7 +88,7 @@ public final class PageIdUtils {
      */
     public static long link(long pageId, int itemId) {
         assert itemId >= 0 && itemId <= MAX_ITEM_ID_NUM : itemId;
-        assert (pageId >> ROTATION_ID_OFFSET) == 0 : IgniteUtils.hexLong(pageId);
+        assert (pageId >> ROTATION_ID_OFFSET) == 0 : HexStringUtils.hexLong(pageId);
 
         return pageId | (((long) itemId) << ROTATION_ID_OFFSET);
     }
