+++ modules/page-memory/src/test/java/org/apache/ignite/internal/pagememory/util/PageIdUtilsTest.java
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Random;
 import org.apache.ignite.internal.pagememory.PageIdAllocator;
-import org.apache.ignite.internal.util.IgniteUtils;
+import org.apache.ignite.internal.util.HexStringUtils;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -131,8 +131,8 @@ public class PageIdUtilsTest {
 
             long pageId = PageIdUtils.pageId(partId, PageIdAllocator.FLAG_DATA, pageNum);
 
-            String msg = "For values [offset=" + IgniteUtils.hexLong(off) + ", fileId=" + IgniteUtils.hexLong(partId)
-                    + ", pageNum=" + IgniteUtils.hexLong(pageNum) + ']';
+            String msg = "For values [offset=" + HexStringUtils.hexLong(off) + ", fileId=" + HexStringUtils.hexLong(partId)
+                    + ", pageNum=" + HexStringUtils.hexLong(pageNum) + ']';
 
             assertEquals(pageId, PageIdUtils.pageId(pageId), msg);
             assertEquals(0, PageIdUtils.itemId(pageId), msg);
