+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/sql/engine/datatypes/varbinary/ItVarBinaryIndexTest.java
@@ -28,7 +28,7 @@ import org.apache.ignite.internal.sql.engine.datatypes.tests.BaseIndexDataTypeTe
 import org.apache.ignite.internal.sql.engine.datatypes.tests.DataTypeTestSpec;
 import org.apache.ignite.internal.sql.engine.datatypes.tests.TestTypeArguments;
 import org.apache.ignite.internal.sql.engine.util.VarBinary;
-import org.apache.ignite.internal.util.IgniteUtils;
+import org.apache.ignite.internal.util.HexStringUtils;
 import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
@@ -121,11 +121,11 @@ public class ItVarBinaryIndexTest extends BaseIndexDataTypeTest<VarBinary> {
                 case LITERAL:
                     return spec.toLiteral(value);
                 case CAST: {
-                    String str = IgniteUtils.toHexString(value.get());
+                    String str = HexStringUtils.toHexString(value.get());
                     return format("x'{}'::VARBINARY", str);
                 }
                 case CAST_WITH_PRECISION: {
-                    String str = IgniteUtils.toHexString(value.get());
+                    String str = HexStringUtils.toHexString(value.get());
                     return format("x'{}'::VARBINARY(8)", str);
                 }
                 default:
