+++ modules/storage-page-memory/src/main/java/org/apache/ignite/internal/storage/pagememory/index/meta/IndexMeta.java
@@ -20,7 +20,7 @@ package org.apache.ignite.internal.storage.pagememory.index.meta;
 import java.util.UUID;
 import org.apache.ignite.internal.tostring.IgniteToStringExclude;
 import org.apache.ignite.internal.tostring.S;
-import org.apache.ignite.internal.util.IgniteUtils;
+import org.apache.ignite.internal.util.HexStringUtils;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -63,6 +63,6 @@ public class IndexMeta extends IndexMetaKey {
 
     @Override
     public String toString() {
-        return S.toString(IndexMeta.class, this, "indexId=", indexId(), "metaPageId", IgniteUtils.hexLong(metaPageId));
+        return S.toString(IndexMeta.class, this, "indexId=", indexId(), "metaPageId", HexStringUtils.hexLong(metaPageId));
     }
 }
